Add HTML-ifying and deployment of webpage
Currently only gets processed via a call to markdown.
Deployment is akin to the one I used in nabijaczleweli/BearLibTerminal.rs:
private key gets decrypted and is used to authenticatedly clone the
gh-pages branch, all new files are added, committed with a message
summarising the commits included with the push to master and pushed.

Tops off and closes #2